reorganize to 1 controller file, remove slackservice as parameter to controller
